+++ src/apis/mypage/fetchBookmarkedProperties.ts
@@ -10,8 +10,13 @@ const fetchBookmarkedProperties = async (
   page: number,
 ): Promise<BookmarkedPropertiesResponse> => {
   try {
-    const response = await axiosInstance.get(`/mypage/histories/bookmarked-properties?page=${page}`);
-    return response.data;
+    // 실제 api 호출
+    // const response = await axiosInstance.get(`/mypage/histories/bookmarked-properties?page=${page}`);
+    // return response.data;
+
+    // mock api 호출
+    const response = await fetch(`/mypage/histories/bookmarked-properties?page=${page}`);
+    return response.json();
   } catch (error) {
     console.error("fetchBookmarkedProperties 에러:", error);
     throw error;
